+++ src/controller/FunctionsController.ts
@@ -1,8 +1,10 @@
 import {ContainerOrchestrationAdapter} from "../adapter/containerOrchestration";
 import {Options} from "../config/Options";
 import {SwarmOrchestrationFactory} from "../factory/SwarmOrchestrationFactory";
+import {Utils} from "./utils";
 
 let containerOrch: ContainerOrchestrationAdapter;
+
 /**
  * @class FunctionsController. Manage BFast::Function instance include
  * deploy and add or remove environment variable(s)
@@ -14,18 +16,10 @@ export class FunctionsController {
             this.options.containerOrchAdapter : new SwarmOrchestrationFactory(this.options);
     }
 
-    private static _checkProjectId(projectId: string): string {
-        if (projectId.length < 1) {
-            throw 'projectId required and can not be empty';
-        } else {
-            return projectId;
-        }
-    }
-
     async deploy(projectId: string = '', force: boolean = false): Promise<any> {
         try {
             return await containerOrch.cloudFunctionsDeploy(
-                FunctionsController._checkProjectId(projectId), force);
+                Utils._checkProjectId(projectId), force);
         } catch (e) {
             throw e.toString();
         }
@@ -34,7 +28,7 @@ export class FunctionsController {
     async envAdd(projectId: string, envs: string[], force: boolean = false): Promise<any> {
         try {
             return await containerOrch.cloudFunctionsAddEnv(
-                FunctionsController._checkProjectId(projectId), envs, force)
+                Utils._checkProjectId(projectId), envs, force);
         } catch (e) {
             throw e.toString();
         }
@@ -43,9 +37,48 @@ export class FunctionsController {
     async envRemove(projectId: string, envs: string[], force: boolean = false): Promise<any> {
         try {
             return await containerOrch.cloudFunctionsRemoveEnv(
-                FunctionsController._checkProjectId(projectId), envs, force)
+                Utils._checkProjectId(projectId), envs, force);
         } catch (e) {
-            throw e.toString();
+            throw e;
         }
     }
+
+    async addDomain(projectId: string, domain: string, force: boolean = false): Promise<any> {
+        try {
+            return await containerOrch.cloudFunctionAddDomain(
+                Utils._checkProjectId(projectId), domain, force);
+        } catch (e) {
+            throw e;
+        }
+    }
+
+    async removeDomain(projectId: string, force: boolean = false): Promise<any> {
+        try {
+            return await containerOrch.cloudFunctionRemoveDomain(
+                Utils._checkProjectId(projectId), force);
+        } catch (e) {
+            throw e;
+        }
+    }
+
+    async faasOn(projectId: string, force: boolean = false): Promise<any> {
+        try {
+            return await containerOrch.cloudFunctionSwitchOn(
+                Utils._checkProjectId(projectId), force
+            );
+        } catch (e) {
+            throw e;
+        }
+    }
+
+    async faasOff(projectId: string, force: boolean = false): Promise<any> {
+        try {
+            return await containerOrch.cloudFunctionSwitchOff(
+                Utils._checkProjectId(projectId), force
+            );
+        } catch (e) {
+            throw e;
+        }
+    }
+
 }