+++ src/router/index.ts
@@ -4,6 +4,8 @@ import {FunctionsRouter} from "./FunctionsRouter";
 import {LandingRouter} from "./LandingRouter";
 import {ProjectRouter} from "./ProjectRouter";
 import {UsersRouter} from "./UsersRouter";
+import {DashboardRouter} from "./DashboardRouter";
+import {DaasRouter} from "./DaasRouter";
 
 export class BFastRouters {
     constructor(private options: Options) {
@@ -14,7 +16,9 @@ export class BFastRouters {
             new FunctionsRouter(this.options),
             new LandingRouter(this.options),
             new ProjectRouter(this.options),
-            new UsersRouter(this.options)
+            new UsersRouter(this.options),
+            new DashboardRouter(this.options),
+            new DaasRouter(this.options),
         ]
     }
 }